fix: file structure, individualized tests, added snapshot testing
