+++ index.js
@@ -1,7 +1,7 @@
 /* eslint-disable prefer-import/prefer-import-over-require */
 
-const withRasterImages = require('./plugins/raster-images');
-const withPlayback = require('./plugins/playback');
-const withSVG = require('./plugins/svg');
+const withRasterImages = require('./plugins/raster-images/raster-images');
+const withPlayback = require('./plugins/playback/playback');
+const withSVG = require('./plugins/svg/svg');
 
 module.exports = { withRasterImages, withPlayback, withSVG };
