+++ plugins/raster-images/raster-images.js
