+++ plugins/svg/svg.js
