+++ plugins/svg/svg.test.js
@@ -1,19 +1,15 @@
-const { withSVG, withRasterImages, withPlayback } = require('./index.js');
+const { withSVG } = require('../../index.js');
 
 const options = {
     dev: true,
     isServer: true,
 };
 
-let config = {
-    module: {
-        rules: [],
-    },
-};
-
 const mockFunction = jest.fn();
 
-afterEach(() => {
+let config;
+
+beforeEach(() => {
     config = {
         module: {
             rules: [],
@@ -23,18 +19,23 @@ afterEach(() => {
     jest.clearAllMocks();
 });
 
+it('should match snapshot for all results', () => {
+    const arrayOfSVGRulesWithInline = [];
+    const arrayOfSVGRulesWithDefault = [];
+
+    arrayOfSVGRulesWithInline.push(withSVG({ inline: true })().webpack(config, options).module.rules);
+    arrayOfSVGRulesWithDefault.push(withSVG()().webpack(config, options).module.rules);
+
+    expect(arrayOfSVGRulesWithInline).toMatchSnapshot();
+    expect(arrayOfSVGRulesWithDefault).toMatchSnapshot();
+});
+
 it('should return a single rule', () => {
     const arrayOfSVGRules = [];
-    const arrayOfImagesRules = [];
-    const arrayOfPlaybackRules = [];
 
     arrayOfSVGRules.push(withSVG()().webpack(config, options).module.rules);
-    arrayOfImagesRules.push(withRasterImages()().webpack(config, options).module.rules);
-    arrayOfPlaybackRules.push(withPlayback()().webpack(config, options).module.rules);
 
     expect(arrayOfSVGRules).toHaveLength(1);
-    expect(arrayOfImagesRules).toHaveLength(1);
-    expect(arrayOfPlaybackRules).toHaveLength(1);
 });
 
 it('should return a function if next config is a function', () => {
@@ -42,11 +43,9 @@ it('should return a function if next config is a function', () => {
         webpack: mockFunction,
     };
 
-    withRasterImages()(nextConfig).webpack(config, options);
-    withPlayback()(nextConfig).webpack(config, options);
     withSVG()(nextConfig).webpack(config, options);
 
-    expect(nextConfig.webpack.mock.calls).toHaveLength(3);
+    expect(nextConfig.webpack.mock.calls).toHaveLength(1);
 });
 
 it('should change loaders in SVG in accordance with inline option', () => {
